Retrained energy regression and particle ID model
The machine learning model for particle ID and energy regression has
been retrained using the HLT TDR samples and a more recent version of
the TICL reconstruction and iterations (CMSSW_11_1_X). The model has not
been changed, but the full data-processing and training has been ported
to TF 2.1. The input and output layout and names are kept the same for
backward compatibility reasons. The performances in terms both of energy
regression and particle ID are considerably better with respect to the
old training. The model has been trained on electrons, photons and
charged hadrons. Neutral hadrons can and will be added in future.